+++ src/main/java/tornadofx/Menu.kt
@@ -88,6 +88,7 @@ fun Menu.radiomenuitem(name: String, toggleGroup: ToggleGroup? = null, keyCombin
     this += radioMenuItem
     return radioMenuItem
 }
+
 fun Menu.checkmenuitem(name: String, keyCombination: KeyCombination?, graphic: Node? = null, op: (CheckMenuItem.() -> Unit)? = null): CheckMenuItem {
     val checkMenuItem = CheckMenuItem(name,graphic)
     keyCombination?.apply { checkMenuItem.accelerator = this }
@@ -95,4 +96,11 @@ fun Menu.checkmenuitem(name: String, keyCombination: KeyCombination?, graphic: N
     op?.let { it.invoke(checkMenuItem) }
     this+= checkMenuItem
     return checkMenuItem
+}
+
+fun Control.contextmenu(op: ContextMenu.() -> Unit): Node {
+    val menu = ContextMenu()
+    op(menu)
+    contextMenu = menu
+    return this
 }